Implement basic application.
The app uses lanterna TUI framework and has the ability to read GNU Cash
sqlite databases as well as OFX (although OFX import is not yet hooked
up).

This implementation allows you to open a database and view the accounts
and transactions within. It includes an implementation of a basic
scrolling table and a tree-like account chooser to filter the
transaction table.